size_t count

+++ OndselSolver/Array.h
@@ -27,8 +27,8 @@ namespace MbD {
 	public:
 		Array() {}
 		Array(std::vector<T> vec) : std::vector<T>(vec) {}
-		Array(int count) : std::vector<T>(count) {}
-		Array(int count, const T& value) : std::vector<T>(count, value) {}
+		Array(size_t count) : std::vector<T>(count) {}
+		Array(size_t count, const T& value) : std::vector<T>(count, value) {}
 		Array(typename std::vector<T>::iterator begin, typename std::vector<T>::iterator end) : std::vector<T>(begin, end) {}
 		Array(std::initializer_list<T> list) : std::vector<T>{ list } {}
 		virtual ~Array() {}
