+++ OndselSolver/EulerParametersDDot.h
@@ -18,8 +18,8 @@ namespace MbD {
     {
         //qEdot aAddot aBddot aCddot 
     public:
-        EulerParametersDDot(int count) : EulerArray<T>(count) {}
-        EulerParametersDDot(int count, const T& value) : EulerArray<T>(count, value) {}
+        EulerParametersDDot(size_t count) : EulerArray<T>(count) {}
+        EulerParametersDDot(size_t count, const T& value) : EulerArray<T>(count, value) {}
         EulerParametersDDot(std::initializer_list<T> list) : EulerArray<T>{ list } {}
 
         //std::shared_ptr<EulerParametersDot<T>> qEdot;
